requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic-settings==2.7.0
python-jose==3.3.0
Jinja2==3.1.4

sqlalchemy==2.0.36
psycopg[binary]==3.2.3

pytest==8.3.4
httpx==0.28.1
